requirements.txt
  # core
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# sentence-transformers for embeddings (local, no external API)
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0

# cyborgdb SDK (official)
cyborgdb>=0.13.0

# web stuff (if you want a tiny UI later)
fastapi>=0.95.0
uvicorn>=0.22.0

# metrics
pytest>=7.0.0
